Better organize classes
